Crate rusoto_serverlessrepo

The AWS Serverless Application Repository makes it easy for developers and enterprises to quickly find and deploy serverless applications in the AWS Cloud. For more information about serverless applications, see Serverless Computing and Applications on the AWS website.

The AWS Serverless Application Repository is deeply integrated with the AWS Lambda console, so that developers of all levels can get started with serverless computing without needing to learn anything new. You can use category keywords to browse for applications such as web and mobile backends, data processing applications, or chatbots. You can also search for applications by name, publisher, or event source. To use an application, you simply choose it, configure any required fields, and deploy it with a few clicks.

You can also easily publish applications, sharing them publicly with the community at large, or privately within your team or across your organization. To publish a serverless application (or app), you can use the AWS Management Console, AWS Command Line Interface (AWS CLI), or AWS SDKs to upload the code. Along with the code, you upload a simple manifest file, also known as the AWS Serverless Application Model (AWS SAM) template. For more information about AWS SAM, see AWS Serverless Application Model (AWS SAM) on the AWS Labs GitHub repository.

The AWS Serverless Application Repository Developer Guide contains more information about the two developer experiences available:

Consuming Applications – Browse for applications and view information about them, including source code and readme files. Also install, configure, and deploy applications of your choosing.

Publishing Applications – Configure and upload applications to make them available to other developers, and publish new versions of applications.

If you're using the service, you're probably looking for ServerlessRepoClient and ServerlessRepo.

Structs

ApplicationDependencySummary
A nested application summary.

ApplicationPolicyStatement
Policy statement applied to the application.

ApplicationSummary
Summary of details about the application.

CreateApplicationRequest

CreateApplicationResponse

CreateApplicationVersionRequest

CreateApplicationVersionResponse

CreateCloudFormationChangeSetRequest

CreateCloudFormationChangeSetResponse

CreateCloudFormationTemplateRequest

CreateCloudFormationTemplateResponse

DeleteApplicationRequest

GetApplicationPolicyRequest

GetApplicationPolicyResponse

GetApplicationRequest

GetApplicationResponse

GetCloudFormationTemplateRequest

GetCloudFormationTemplateResponse

ListApplicationDependenciesRequest

ListApplicationDependenciesResponse

ListApplicationVersionsRequest

ListApplicationVersionsResponse

ListApplicationsRequest

ListApplicationsResponse

ParameterDefinition
Parameters supported by the application.

ParameterValue
Parameter value of the application.

PutApplicationPolicyRequest

PutApplicationPolicyResponse

RollbackConfiguration
This property corresponds to the AWS CloudFormation RollbackConfiguration Data Type.

RollbackTrigger
This property corresponds to the AWS CloudFormation RollbackTrigger Data Type.

ServerlessRepoClient
A client for the AWSServerlessApplicationRepository API.

Tag
This property corresponds to the AWS CloudFormation Tag Data Type.

UpdateApplicationRequest

UpdateApplicationResponse

Version
Application version details.

VersionSummary
An application version summary.

Enums

CreateApplicationError
Errors returned by CreateApplication

CreateApplicationVersionError
Errors returned by CreateApplicationVersion

CreateCloudFormationChangeSetError
Errors returned by CreateCloudFormationChangeSet

CreateCloudFormationTemplateError
Errors returned by CreateCloudFormationTemplate

DeleteApplicationError
Errors returned by DeleteApplication

GetApplicationError
Errors returned by GetApplication

GetApplicationPolicyError
Errors returned by GetApplicationPolicy

GetCloudFormationTemplateError
Errors returned by GetCloudFormationTemplate

ListApplicationDependenciesError
Errors returned by ListApplicationDependencies

ListApplicationVersionsError
Errors returned by ListApplicationVersions

ListApplicationsError
Errors returned by ListApplications

PutApplicationPolicyError
Errors returned by PutApplicationPolicy

UpdateApplicationError
Errors returned by UpdateApplication

Traits

ServerlessRepo
Trait representing the capabilities of the AWSServerlessApplicationRepository API. AWSServerlessApplicationRepository clients implement this trait.