Struct rusoto_serverlessrepo::ParameterDefinition

pub struct ParameterDefinition {
    pub allowed_pattern: Option<String>,
    pub allowed_values: Option<Vec<String>>,
    pub constraint_description: Option<String>,
    pub default_value: Option<String>,
    pub description: Option<String>,
    pub max_length: Option<i64>,
    pub max_value: Option<i64>,
    pub min_length: Option<i64>,
    pub min_value: Option<i64>,
    pub name: Option<String>,
    pub no_echo: Option<bool>,
    pub referenced_by_resources: Option<Vec<String>>,
    pub type_: Option<String>,
}

Parameters supported by the application.

Fields
allowed_pattern: Option<String>

A regular expression that represents the patterns to allow for String types.

allowed_values: Option<Vec<String>>

Array containing the list of values allowed for the parameter.

constraint_description: Option<String>

A string that explains a constraint when the constraint is violated. For example, without a constraint description,\n a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user\n specifies an invalid value:\n\n Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+ \n \nBy adding a constraint description, such as "must contain only uppercase and lowercase letters, and numbers," you can display\n the following customized error message:\n\n Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

default_value: Option<String>

A value of the appropriate type for the template to use if no value is specified when a stack is created.\n If you define constraints for the parameter, you must specify a value that adheres to those constraints.

description: Option<String>

A string of up to 4,000 characters that describes the parameter.

max_length: Option<i64>

An integer value that determines the largest number of characters you want to allow for String types.

max_value: Option<i64>

A numeric value that determines the largest numeric value you want to allow for Number types.

min_length: Option<i64>

An integer value that determines the smallest number of characters you want to allow for String types.

min_value: Option<i64>

A numeric value that determines the smallest numeric value you want to allow for Number types.

name: Option<String>

The name of the parameter.

no_echo: Option<bool>

Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the\n value to true, the parameter value is masked with asterisks (*****).

referenced_by_resources: Option<Vec<String>>

A list of SAM resources that use this parameter.

type_: Option<String>

The type of the parameter.\nValid values: String | Number | List | CommaDelimitedList \n \n\n String : A literal string.\nFor example, users could specify "MyUserName" .\n\n Number : An integer or float. AWS CloudFormation validates the parameter value as a number; however, when you use the\n parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.\nFor example, users could specify "8888" .\n\n List : An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers; however, when\n you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.\nFor example, users could specify "80,20", and a Ref results in ["80","20"] .\n\n CommaDelimitedList : An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas.\n Also, each member string is space-trimmed.\nFor example, users could specify "test,dev,prod", and a Ref results in ["test","dev","prod"] .