Trait rusoto_serverlessrepo::ServerlessRepo

pub trait ServerlessRepo {
    fn create_application(
        &self, 
        input: CreateApplicationRequest
    ) -> RusotoFuture<CreateApplicationResponse, CreateApplicationError>;

    fn create_application_version(
        &self, 
        input: CreateApplicationVersionRequest
    ) -> RusotoFuture<CreateApplicationVersionResponse, CreateApplicationVersionError>;

    fn create_cloud_formation_change_set(
        &self, 
        input: CreateCloudFormationChangeSetRequest
    ) -> RusotoFuture<CreateCloudFormationChangeSetResponse, CreateCloudFormationChangeSetError>;

    fn create_cloud_formation_template(
        &self, 
        input: CreateCloudFormationTemplateRequest
    ) -> RusotoFuture<CreateCloudFormationTemplateResponse, CreateCloudFormationTemplateError>;

    fn delete_application(
        &self, 
        input: DeleteApplicationRequest
    ) -> RusotoFuture<(), DeleteApplicationError>;

    fn get_application(
        &self, 
        input: GetApplicationRequest
    ) -> RusotoFuture<GetApplicationResponse, GetApplicationError>;

    fn get_application_policy(
        &self, 
        input: GetApplicationPolicyRequest
    ) -> RusotoFuture<GetApplicationPolicyResponse, GetApplicationPolicyError>;

    fn get_cloud_formation_template(
        &self, 
        input: GetCloudFormationTemplateRequest
    ) -> RusotoFuture<GetCloudFormationTemplateResponse, GetCloudFormationTemplateError>;

    fn list_application_dependencies(
        &self, 
        input: ListApplicationDependenciesRequest
    ) -> RusotoFuture<ListApplicationDependenciesResponse, ListApplicationDependenciesError>;

    fn list_application_versions(
        &self, 
        input: ListApplicationVersionsRequest
    ) -> RusotoFuture<ListApplicationVersionsResponse, ListApplicationVersionsError>;

    fn list_applications(
        &self, 
        input: ListApplicationsRequest
    ) -> RusotoFuture<ListApplicationsResponse, ListApplicationsError>;

    fn put_application_policy(
        &self, 
        input: PutApplicationPolicyRequest
    ) -> RusotoFuture<PutApplicationPolicyResponse, PutApplicationPolicyError>;

    fn update_application(
        &self, 
        input: UpdateApplicationRequest
    ) -> RusotoFuture<UpdateApplicationResponse, UpdateApplicationError>;
}

Trait representing the capabilities of the AWSServerlessApplicationRepository API. AWSServerlessApplicationRepository clients implement this trait.

Required methods

fn create_application(
&self,
input: CreateApplicationRequest
) -> RusotoFuture<CreateApplicationResponse, CreateApplicationError>

Creates an application, optionally including an AWS SAM file to create the first application version in the same call.

fn create_application_version(
&self,
input: CreateApplicationVersionRequest
) -> RusotoFuture<CreateApplicationVersionResponse, CreateApplicationVersionError>

Creates an application version.

fn create_cloud_formation_change_set(
&self,
input: CreateCloudFormationChangeSetRequest
) -> RusotoFuture<CreateCloudFormationChangeSetResponse, CreateCloudFormationChangeSetError>

Creates an AWS CloudFormation change set for the given application.

fn create_cloud_formation_template(
&self,
input: CreateCloudFormationTemplateRequest
) -> RusotoFuture<CreateCloudFormationTemplateResponse, CreateCloudFormationTemplateError>

Creates an AWS CloudFormation template.

fn delete_application(
&self,
input: DeleteApplicationRequest
) -> RusotoFuture<(), DeleteApplicationError>

Deletes the specified application.

fn get_application(
&self,
input: GetApplicationRequest
) -> RusotoFuture<GetApplicationResponse, GetApplicationError>

Gets the specified application.

fn get_application_policy(
&self,
input: GetApplicationPolicyRequest
) -> RusotoFuture<GetApplicationPolicyResponse, GetApplicationPolicyError>

Retrieves the policy for the application.

fn get_cloud_formation_template(
&self,
input: GetCloudFormationTemplateRequest
) -> RusotoFuture<GetCloudFormationTemplateResponse, GetCloudFormationTemplateError>

Gets the specified AWS CloudFormation template.

fn list_application_dependencies(
&self,
input: ListApplicationDependenciesRequest
) -> RusotoFuture<ListApplicationDependenciesResponse, ListApplicationDependenciesError>

Retrieves the list of applications nested in the containing application.

fn list_application_versions(
&self,
input: ListApplicationVersionsRequest
) -> RusotoFuture<ListApplicationVersionsResponse, ListApplicationVersionsError>

Lists versions for the specified application.

fn list_applications(
&self,
input: ListApplicationsRequest
) -> RusotoFuture<ListApplicationsResponse, ListApplicationsError>

Lists applications owned by the requester.

fn put_application_policy(
&self,
input: PutApplicationPolicyRequest
) -> RusotoFuture<PutApplicationPolicyResponse, PutApplicationPolicyError>

Sets the permission policy for an application. For the list of actions supported for this operation, see Application Permissions .

fn update_application(
&self,
input: UpdateApplicationRequest
) -> RusotoFuture<UpdateApplicationResponse, UpdateApplicationError>

Updates the specified application.

Loading content...