Struct rusoto_serverlessrepo::ServerlessRepoClient

pub struct ServerlessRepoClient { /* fields omitted */ }

A client for the AWSServerlessApplicationRepository API.

Methods

impl ServerlessRepoClient

pub fn new(region: Region) -> ServerlessRepoClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ServerlessRepoClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,